backend/pyproject.toml
[project]
name = "pydantic-ai-neo4j-backend"
version = "0.1.0"
description = "FastAPI backend for news chat agent with Pydantic AI and Neo4j"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic-ai>=1.0.0",
    "neo4j==5.14.1",
    "python-dotenv==1.0.0",
    "pydantic>=2.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

